+++ ewb-dig-ag-site/src/App.js
@@ -1,10 +1,10 @@
 import "./App.css";
 import NavContent from "./components/NavContent";
 import Home from "./components/Home";
-import Overview from "./components/Overview";
+import NLB from "./components/NLB";
 import Team from "./components/Team";
 import Contact from "./components/Contact";
-import Process from "./components/Process";
+import Solution from "./components/Solution";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import {} from "./components/shared/Section.styles";
@@ -16,8 +16,8 @@ function App() {
         <NavContent></NavContent>
         <Routes>
           <Route path="/" element={<Home />} />
-          <Route path="overview" element={<Overview />} />
-          <Route path="process" element={<Process />} />
+          <Route path="nlb" element={<NLB />} />
+          <Route path="solution" element={<Solution />} />
           <Route path="team" element={<Team />} />
           <Route path="contact" element={<Contact />} />
         </Routes>
